+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change Log
 
 ## [Unreleased]
-- Added "observer" attribute to @property annotation
-- Added "computed" attribute to @property annotation
-- Fix bug where use of `Polymer.mixinBehaviors` with `@customElement` would throw an exception relating to the definition of the element's "properties" property.
+
+## [0.1.2] - 2018-01-01
+- Added `observer` and `computed` attribute to `@property` decorator.
+- Added `@listen` decorator and `DeclarativeEventListeners` mixin.
+- Fixed bug where use of `Polymer.mixinBehaviors` with `@customElement` would throw an exception relating to the definition of the element's "properties" property.
 
 ## [0.1.1] - 2017-09-28
 - Metadata Reflection polyfill is no longer a bower dependency.